Rename and relocate classes
